Definitive migration to react, deletion of the web folder
