--- Recife-Tecnologia/src/css/geral/geral.css
@@ -1,12 +1,0 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-}
-
-
-
-
-
-
-
