--- Recife-Tecnologia/src/pages/RecoveryPass/RecTeste.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const RecTeste = () => {
-  return (
-    <div className="teste-background">
-      <h1>hello world</h1>
-    </div>
-  );
-};
-
-export default RecTeste;
