--- Recife-Tecnologia/src/pages/RecoveryPass/style.css
@@ -1,3 +1,0 @@
-body {
-  color: pink;
-}
