+++ Recife-Tecnologia/src/routes/AppRoutes.jsx
@@ -1,10 +1,8 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Login from '../pages/Login';
-import RecoveryPass from '../pages/RecoveryPass';
-import RecTeste from '../pages/RecoveryPass/RecTeste';
 import withBodyStyles from '../components/StylesBody';
 import Dashboard from '../pages/Dashboard';
-import Relatorio from '../pages/Relatório';
+import Relatorio from '../pages/Report';
 
 const AppRoutes = () => {
   console.log('Passou por aqui');
@@ -14,7 +12,6 @@ const AppRoutes = () => {
         <Route path="/" element={<Login />} />
         <Route path="/Dashboard" element={<Dashboard />} />
         <Route path="/Relatorio" element={<Relatorio />} />
-
       </Routes>
     </BrowserRouter>
   );
