--- web/css/dashboard/dashboard.css
@@ -1,132 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-.header-bg {
-  background-color: var(--cor-P1);
-  padding: 30px 0 30px 0;
-}
-
-.container-imgsearch {
-  border-top-left-radius: 10px;
-  border-bottom-left-radius: 10px;
-  background-color: var(--cor-P1);
-  width: 30px;
-  padding-left: 5px;
-  display: flex;
-  align-items: center;
-}
-
-.header-container .search {
-  width: 70px;
-  height: 30px;
-  border-top-right-radius: 10px;
-  border-bottom-right-radius: 10px;
-  border: none;
-  background-color: var(--cor-P1);
-  color: var(--cor-B0);
-  font-size: 8pt;
-  transition: all 1s;
-  display: inline-flex;
-  text-align: start;
-}
-
-.header-container .search:focus {
-  background-color: var(--cor-B0);
-  color: var(--cor-P1);
-  width: 200px;
-  transition: all 1s;
-  outline: none;
-  font-size: 10pt;
-  margin-left: 5px;
-  padding-left: 5px;
-}
-
-.container-search {
-  width: 200px;
-  display: inline-flex;
-}
-
-.container-dashboard{
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.header-container {
-  max-width: 1200px;
-  margin: 0 auto;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.search-bar {
-  display: flex;
-  align-items: center;
-}
-
-input[type="text"] {
-  padding: 5px;
-  border: 1px solid #ccc;
-}
-
-
-
-.user-menu ul {
-  list-style: none;
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.user-menu a {
-  color: var(--cor-B0);
-  text-decoration: none;
-  position: relative;
-}
-
-.user-menu a::after {
-  content: "";
-  display: block;
-  height: 2px;
-  width: 0px;
-  background: var(--COR-L0);
-  margin-top: 4px;
-  transition: 0.3s;
-  position: absolute;
-}
-
-.user-menu a:hover::after {
-  width: 90%;
-}
-
-.user-menu li {
-  margin-right: 10px;
-  cursor: pointer;
-}
-
-.user-menu .google-icon {
-  margin-right: 20px;
-}
-
-
-.menu-list {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.menu-list li {
-  padding: 10px;
-  border-bottom: 1px solid var(--cor-B0);
-}
-
-#mobile-menu {
-  display: none;
-}
