--- web/css/loginScreen/LoginScreen.css
@@ -1,122 +1,0 @@
-
-.login-container {
-    background-color: var(--cor-P0);
-    flex-direction: column;
-    align-items: center;
-    max-width: 100%;
-    border-radius: 20px;
-    border: 1px solid var(--cor-P0);
-    box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25) inset;
-    overflow: hidden;
-    margin: 20px 30px 20px 30px;
-    padding: 20px;
-    display: flex;
-}
-
-.login-content {
-    
-    grid-template-columns: 1fr 1fr;
-    background-color: var(--cor-B0);
-    box-shadow: 2px 2px 2px 0px rgba(194, 194, 194, 0.25);
-    text-align: center;
-    padding: 20px;
-    gap: 80px;
-    border-radius: 30px;
-    display: grid;
-    justify-content: center;
-    
-}
-
-
-.login-content_img {
-    border-radius: 20px;
-    box-shadow: 2px 2px 2px 0px rgba(194, 194, 194, 0.25);
-    border: 1px solid #B6AAAA;
-    margin: 180px 60px 180px 60px;
-    max-width: 100%;
-    max-height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-}
-
-
-.login-form-container {
-    
-    flex-direction: column;
-    border-radius: 20px;
-    border: 1px solid #B6AAAA;
-    background: var(--B1, #FFF);
-    box-shadow: 2px 2px 2px 0px rgba(194, 194, 194, 0.25);
-    margin: 120px 20px 120px 20px;
-    max-width: 360px;
-    max-height: 480px;
-    align-items: center;
-    display: flex;
-    padding: 20px;
-}
-
-
-.login-content_form_img {
-    margin: 10px 70px 10px 70px;
-}
-
-
-.login-form_content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 20px;   
-}
-
-.login-form_content a#recoveryPass {
-    color: var(--COR-C1);
-    text-decoration: none;
-    margin: 10px 0px 10px 185px;
-    flex-wrap: nowrap;
-    white-space: nowrap;
-}
-
-.login-form_content a#recoveryPass:hover {
-color: var(--COR-L1);
-transition: 0.3s;
-}
-
-.login-form_content button#enterAccount {
-    margin: 10px 0px 10px 0px;
-}
-
-.login-form_content a#enterAccount:hover {
-    background-color: var(--COR-L0);
-    color: var(--cor-B1);
-    transition: 0.5s;
-}
-
-
-
-.login-form_content label {
-    margin: 5px 0px;
-    margin-right: 160px;
-}
-
-
-.login-form_input {
-    width: 100%;
-    height: 40px;
-    flex-shrink: 0;
-    margin: 5px 0px;
-    padding: 0px 10px;
-    border-radius: 8px;
-    border: 1px solid #B6AAAA;
-    background: var(--B0, #FFF);
-    box-shadow: 2px 2px 2px 0px rgba(194, 194, 194, 0.25);
-    text-align: center;
-}
-
-
-.login-form_input:focus {
-    border: 1px solid var(--COR-L0);
-    outline: none;
-}
-
