--- web/css/responsive/LoginResponse.css
@@ -1,113 +1,0 @@
-
-@media (max-width: 1920px) {
-    .login-container {
-        padding-top: 10px;
-        margin-bottom: 10px;
-    }
-    .login-content {
-        max-height: 600PX;
-    }
-    .login-form-container {
-        margin-top: 60px;
-        margin-bottom: 30px;
-    }
-    .login-content_img {
-        margin-top: 130px;
-    }
-}
-@media (max-width: 1366px) {
-    .login-container {
-        padding-top: 10px;
-    }
-}
-
-
-@media (max-width: 1366px) {
-    .login-container {
-        padding-top: 10px;
-    }
-}
-
-@media (max-width: 1000px) {
-    #body-login {
-        background-color: white;
-    }
-    .login-container {
-        background:white;
-        border: none;
-        box-shadow: none;
-    }
-    .login-content {
-        grid-template-columns: 1fr;
-        gap: 0px;
-        margin: 30px 10px 30px 10px;
-        border: 1px solid #B6AAAA;;
-        
-    }
-    .login-content_img {
-        display: none;
-    }
-    .login-form-container {
-        margin: 0px;
-        max-width: 100%;
-        max-height: 100%;
-        border: none;
-    }
-    .login-content_form_img {
-        margin: 10px 70px 10px 70px;
-
-    }
-    .login-form_content {
-        padding: 0px;
-    }
-    .login-form_content a#recoveryPass {
-        margin-top: 5px;
-    }
-    .login-form_content button#enterAccount {
-        margin-top: 10px;
-    }
-}
-
-@media (max-width: 500px) {
-    body {
-        background-color: white;
-    }
-    .login-form-container {
-        max-width: 250px;
-        width: auto;
-        border: none;
-    }
-    .login-container {
-        background:white;
-        border: none;
-        box-shadow: none;
-    }
-    
-    .login-content {
-        
-        border: 1px solid #B6AAAA;;
-    }
-    .login-content_form_img {
-        margin: 10px 10px 10px 10px;
-    }
-    .login-form_content {
-        max-width: 200px;
-    }
-    .login-form_content label {
-        font-size: 16px;
-        flex-wrap: nowrap;
-        white-space: nowrap;
-        margin-left: 70px;
-    }
-    .login-form_content a#recoveryPass {
-        margin: 5px 30px;
-    }
-}
-
-@media (max-width: 320px) {
-    .login-content {
-        margin: 60px 0px 60px 0px;
-        width: 200px;
-        
-    }
-}
