--- web/css/style.css
@@ -1,11 +1,0 @@
-@import './utility/geral.css';
-@import './utility/colors.css';
-@import './utility/fonts.css';
-@import './utility/buttons.css';
-
-@import './dashboard/dashboard.css';
-@import './loginScreen/LoginScreen.css';
-@import './responsive/LoginResponse.css';
-@import './responsive/DashboardResponse.css';
-
-
