--- web/css/utility/buttons.css
@@ -1,21 +1,0 @@
-.button-m {
-    color: white;
-    background-color: var(--COR-L0);
-    border: none;
-    border-radius: 20px;
-    width: 140px;
-    height: 42px;
-    font-size: 18px;
-    font-weight: 400;
-    font-family: 'Montserrat', sans-serif;
-    cursor: pointer;
-    transition: 0.3s;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.button-m:hover {
-    color: var(--COR-L1);
-    transition: 0.3s;
-}
