--- web/css/utility/colors.css
@@ -1,10 +1,0 @@
-:root {
-    --cor-P0: #070707;
-    --cor-P1: #000000;
-    --cor-P2: #101010;
-    --cor-P3: #424242;
-    --cor-B0: #FFFFFF;
-    --COR-L0: #FF7707;
-    --COR-C0: #B8B8B8;
-    --COR-C1: #A8A8A8;
-}
