--- web/css/utility/geral.css
@@ -1,59 +1,0 @@
-.container {
-  box-sizing: border-box;
-  width: 1200px;
-  padding-left: 20px;
-  padding-right: 20px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-#body-login {
-  margin: 0;
-  font-family: Montserrat, sans-serif;
-  background-color: var(--cor-P0);
-  display: flex;
-  justify-content: center; 
-  align-items: center; 
-  min-height: 100vh; 
-}
-#body-dashboard{
-  margin: 0px;
-  font-family: Montserrat, sans-serif;
-  background-color: var(--cor-P0); 
-}
-
-h1,
-h2,
-h3,
-h4,
-ul,
-p,
-blockquote,
-dd {
-  margin: 0px;
-}
-
-ul {
-  padding: 0px;
-  list-style: none;
-}
-
-
-a {
-  text-decoration: none;
-  color: inherit;
-}
-
-img {
-  max-width: 100%;
-  display: block;
-}
-
-img[src$=".jpg"] {
-  border-radius: 4px;
-}
-input[type="search"]:focus::-webkit-search-cancel-button {
-  -webkit-appearance: none;
-  appearance: none;
-  display: none;
-}
