--- web/dashboard.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./css/style.css">
-    <link rel="stylesheet"
-        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
-    <title>Dashboard</title>
-</head>
-<body id="body-dashboard">
-    <header class="header-bg">
-        <div class="header-container">
-            <div class="container-search">
-                <div class="container-imgsearch"><img src="./img/Dashboard/find.png" alt="search"></div>
-                <input type="search" name="search" placeholder="Pesquisar..." class="search" />
-            </div>
-            <nav class="user-menu font-1-m" id="nav-menu">
-                <span id="hamburger">
-                    
-                
-                    <div></div>
-                </span>
-                <ul id="ul-menu" role="menu">
-                    <li><a href="">Dashboard</a></li>
-                    <li><a href="">Relatórios</a></li>
-                    <li><a  id="hora-atual1">Hora</a></li>
-                    <li><a  id="data-atual1">Data</a></li>
-                    <li><a href="">Nome Cliente</a></li>
-                    <li><a href=""><img src="./img/Dashboard/exitIcon.svg" alt="Google Icon" width="32" height="32"></a></li>
-                </ul>
-            </nav>
-            
-            <div id="mobile-menu">
-                <ul>
-                    <li><a href="">Dashboard</a></li>
-                    <li><a href="">Relatórios</a></li>
-                    <li><a  id="hora-atual2">Hora</a></li>
-                    <li><a  id="data-atual2">Data</a></li>
-                    <li><a href="">Nome Cliente</a></li>
-                    <li><a href="">Minha conta</a></li>
-                </ul>
-            </div>
-        </div>
-    </header>
-    <div class="container-dashboard">
-    <iframe title="Dashboard" width="1500" height="1600" src="https://app.powerbi.com/reportEmbed?reportId=2b32a6bc-b3f3-4ed7-b33b-cdbeb7b6dc22&autoAuth=true&ctid=2009dfae-df11-49c7-804d-fda8d5cd9865" frameborder="0" allowFullScreen="true"></iframe>
-    </div>
-
-    <script type="module" src="./script/script.js"></script>
-</body>
-</html>
