--- web/index.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./css/style.css">
-    <title>Tela de login</title>
-</head>
-
-<body id="body-login">
-    <div class="login-container">
-        <div class="login-content">
-            <img src="./img/LoginScreen/WomanPickupRelatories.jpg" class="login-content_img" alt="Mulher selecionando relatórios, imagem ilustrativa para compor a tela inicial" width="560px" height="360px">
-            <div class="login-form-container">
-                <form id="loginForm">
-                    <img src="./img/LoginScreen/Logo.png" alt="Logo da empresa Recife Tecnologia" class="login-content_form_img">
-                    <div class="login-form_content">
-                        <label for="login" class="font-1-m">Insira seu email:</label>
-                        <input type="text" name="login" id="login" placeholder="seuemail@.com" class="login-form_input" autocomplete="username">
-                        <label for="password" class="font-1-m">Insira sua senha:</label>
-                        <input type="password" name="password" id="password" placeholder="********" class="login-form_input" autocomplete="current-password">
-                        <p id="userError" class="error-message"></p>
-                        <p id="passwordError" class="error-message"></p>
-                        <a href="./recoveryPass.html" class="font-1-pp" id="recoveryPass">Esqueceu a senha?</a>
-                        <button type="submit" class="font-1-m button-m" id="enterAccount">Login</button>
-                    </div>
-                </form>
-            </div>
-        </div>
-    </div>
-    <script type="module" src="./script/script.js"></script>
-</body>
