--- web/script/script.js
@@ -1,7 +1,0 @@
-import InitvalidationLogin from './modules/validationLogin.js';
-import InitdataLoad from './modules/dataLoad.js';
-import InitmobileMenu from './modules/mobileMenu.js';
-
-InitvalidationLogin();
-InitdataLoad();
-InitmobileMenu();
