--- web/testarea/indexTeste.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./testeCss.css">
-    <link rel="stylesheet"
-        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
-    <title>Dashboard</title>
-</head>
-<body id="body-dashboard">
-    <header class="header-bg">
-        <div class="header-container">
-            <div class="container-search">
-                <div class="container-imgsearch"><img src="../img/Dashboard/find.png" alt="search"></div>
-                <input type="search" name="search" placeholder="Pesquisar..." class="search" />
-            </div>
-            <nav class="user-menu font-1-m" id="nav-menu">
-                <span id="hamburger">
-                    
-                
-                    <div></div>
-                </span>
-                <ul id="ul-menu" role="menu">
-                    <li><a href="">Dashboard</a></li>
-                    <li><a href="">Relatórios</a></li>
-                    <li><a href="" id="hora-atual">Hora</a></li>
-                    <li><a href="" id="data-atual">Data</a></li>
-                    <li><a href="">Nome Cliente</a></li>
-                    <li><a href=""><img src="../img/Dashboard/exitIcon.svg" alt="Google Icon" width="32" height="32"></a></li>
-                </ul>
-            </nav>
-            
-            <div id="mobile-menu">
-                <ul>
-                    <li><a href="">Dashboard</a></li>
-                    <li><a href="">Relatórios</a></li>
-                    <li><a href="">Hora</a></li>
-                    <li><a href="">Data</a></li>
-                    <li><a href="">Nome Cliente</a></li>
-                    <li><a href="">Minha conta</a></li>
-                </ul>
-            </div>
-        </div>
-    </header>
-
-    <main>
-       
-    </main>
-
-    <script type="module" src="./scriptTeste.js"></script>
-</body>
-</html>
