--- web/testarea/scriptTeste.js
@@ -1,7 +1,0 @@
-const hamburger = document.getElementById('hamburger');
-const mobileMenu = document.getElementById('mobile-menu');
-
-hamburger.addEventListener('click', () => {
-    hamburger.classList.toggle('closed');
-    mobileMenu.classList.toggle('active');
-});
