--- web/testarea/testeCss.css
@@ -1,231 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-.header-bg {
-    background-color: pink;
-    padding: 30px 0 30px 0;
-}
-
-.container-imgsearch {
-    border-top-left-radius: 10px;
-    border-bottom-left-radius: 10px;
-    background-color: black;
-    width: 30px;
-    padding-left: 5px;
-    display: flex;
-    align-items: center;
-}
-
-.header-container .search {
-    width: 70px;
-    height: 30px;
-    border-top-right-radius: 10px;
-    border-bottom-right-radius: 10px;
-    border: none;
-    background-color: black;
-    color: white;
-    display: inline-flex;
-    text-align: start;
-}
-
-.header-container .search:focus {
-    background-color: white;
-    color: black;
-    width: 200px;
-    transition: all 1s;
-    outline: none;
-}
-
-.container-search {
-    width: 200px;
-    display: inline-flex;
-}
-
-.header-container {
-    max-width: 1200px;
-    margin: 0 auto;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.search-bar {
-    display: flex;
-    align-items: center;
-}
-
-input[type="text"] {
-    padding: 5px;
-    border: 1px solid #ccc;
-}
-
-
-
-.user-menu ul {
-    list-style: none;
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
-
-.user-menu a {
-    color: black;
-    text-decoration: none;
-    position: relative;
-}
-
-.user-menu a::after {
-    content: "";
-    display: block;
-    height: 2px;
-    width: 0px;
-    background: orange;
-    margin-top: 4px;
-    transition: 0.3s;
-    position: absolute;
-}
-
-.user-menu a:hover::after {
-    width: 90%;
-}
-
-.user-menu li {
-    margin-right: 10px;
-    cursor: pointer;
-}
-
-.user-menu .google-icon {
-    margin-right: 20px;
-}
-
-
-.menu-list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-.menu-list li {
-    padding: 10px;
-    border-bottom: 1px solid white;
-}
-
-#mobile-menu {
-    display: none;
-}
-  
-
-@media (max-width: 600px) {
-    #ul-menu {
-        display: none;
-        position: absolute;
-        top: 70px;
-        right: 20px; 
-        background: #e7e7e7;
-        transition: 0.6s;
-        z-index: 1000;
-    }
-
-    #nav-menu.active #ul-menu {
-        display: block;
-    }
-
-    #hamburger {
-        cursor: pointer;
-        display: block;
-        width: 30px;
-        height: 30px;
-        position: relative;
-    }
-    
-    #hamburger .bar {
-        content: '';
-        display: block;
-        width: 20px;
-        height: 2px;
-        background: currentColor;
-        margin-top: 5px;
-        transition: 0.3s;
-        position: relative;
-    }
-    
-    #hamburger::after,
-    #hamburger::before,
-    #hamburger div {
-      content: '';
-      display: block;
-      width: 20px;
-      height: 2px;
-      background: currentColor;
-      margin-top: 5px;
-      transition: 0.3s;
-      position: relative;
-    }
-
-    #hamburger.closed::before {
-        transform: rotate(45deg);
-        margin-top: 20px;
-        top: 0px;
-        transform-origin: center;
-    }
-    
-    #hamburger.closed::after {
-        transform: rotate(-45deg);
-        bottom: 7px;
-        transform-origin: center;
-    }
-    
-    #hamburger.closed div {
-        display: none;
-    }
-
-    
-   
-    #mobile-menu {
-        display: none;
-        position: absolute;
-        top: 70px;
-        right: 20px;
-        background: #e7e7e7;
-        transition: 0.6s;
-        z-index: 1000;
-        width: 200px;
-        box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
-    }
-    
-    #mobile-menu.active {
-        display: block;
-    }
-    
-    #mobile-menu ul {
-        list-style: none;
-        padding: 0;
-        margin: 0;
-    }
-    
-    #mobile-menu li {
-        margin: 10px 0;
-    }
-    
-    #mobile-menu a {
-        display: block;
-        padding: 10px 20px;
-        color: #333;
-        text-decoration: none;
-        transition: background-color 0.3s;
-    }
-    
-    #mobile-menu a:hover {
-        background-color: #ddd;
-    }
-    
-
-}
-
-
-
-
-
