--- web/users.json
@@ -1,6 +1,0 @@
-{
-    "users": [{
-        "email": "squad7@gmail.com",
-        "password": "teste123"
-    }]
-}
